requirements.txt
Flask
pdfplumber
python-docx
fuzzywuzzy
python-Levenshtein